+++ src/auto/utils/aws_context.py
@@ -12,15 +12,15 @@ class AwsContext:
     nor on the ~/.aws credentials and config files (nor on the AWS_SHARED_CREDENTIALS_FILE
     and AWS_CONFIG_FILE environment variables).
 
-    EITHER a specific path to the ~/.aws credentials directory MUST be specified, which
-    will will setup the context to refer to the credentials and config file(s) there;
+    EITHER a specific path to the ~/.aws credentials directory MUST be specified,
+    which will setup the context to refer to the credentials and config file(s) there;
     OR, specific AWS access key ID and associated secret access key (and region) values
     MUST be specified; the latter taking precedence over the former. Usage like this:
 
         aws = AwsContext(your_aws_credentials_directory_or_access_key_id_and_secret_access_key)
         with aws.establish_credentials() as credentials:
             do_something_with_boto3()
-            # if desired reference/use credentials values ...
+            # if desired reference/use credentials values like so:
             aws_access_key_id = credentials.access_key_id
             aws_secret_access_key = credentials.secret_access_key
             aws_region = credentials.region
@@ -29,8 +29,10 @@ class AwsContext:
     """
 
     def __init__(self,
-                 aws_credentials_dir: str,
-                 aws_access_key_id: str = None, aws_secret_access_key: str = None, aws_region: str = None,
+                 aws_credentials_dir: str = None,
+                 aws_access_key_id: str = None,
+                 aws_secret_access_key: str = None,
+                 aws_region: str = None,
                  aws_session_token: str = None) -> None:
         """
         Constructor which stores the given AWS credentials directory, and AWS access key ID
@@ -41,6 +43,7 @@ class AwsContext:
         :param aws_access_key_id: AWS credentials access key ID.
         :param aws_secret_access_key: AWS credentials secret access key.
         :param aws_region: AWS credentials region.
+        :param aws_session_token: AWS session token. NOTE: Not yet tested at all.
         """
         self._aws_access_key_id = aws_access_key_id
         self._aws_secret_access_key = aws_secret_access_key
@@ -51,8 +54,11 @@ class AwsContext:
 
     class Credentials:
         def __init__(self,
+                     credentials_dir: str, credentials_dir_symlink_target: str,
                      access_key_id: str, secret_access_key: str, region: str,
                      account_number: str, user_arn: str) -> None:
+            self.credentials_dir = credentials_dir
+            self.credentials_dir_symlink_target = credentials_dir_symlink_target
             self.access_key_id = access_key_id
             self.secret_access_key = secret_access_key
             self.region = region
@@ -70,25 +76,21 @@ class AwsContext:
         manager context) blow away any pertinent AWS credentials related environment variables,
         and set them appropriately based on given EXPLICITLY specified credentials information.
 
-        :param display: If True then print summary of AWS credentials.
+        :param display: If True then PRINT summary of AWS credentials.
         :param show: If True and display True show in plaintext sensitive info for AWS credentials summary.
-        :return: Yields named tuple with: access_key_id, secret_access_key, region, account_number, user_arn.
+        :return: Yields populated (nested class) Credentials object.
         """
 
-        # TODO: Should we require all credentials, INCLUDING region, to come from EITHER
-        # given arguments (i.e. command-line, ultimately) XOR from given AWS credentials
-        # directory? I.e. so as not to split between them which may create some confusion.
-
         def unset_environ(environment_variables: list) -> dict:
-            saved_environ = {}
+            saved_environment_variables = {}
             for environment_variable in environment_variables:
-                saved_environ[environment_variable] = os.environ.pop(environment_variable, None)
+                saved_environment_variables[environment_variable] = os.environ.pop(environment_variable, None)
                 if environment_variable.endswith("_FILE"):
                     os.environ[environment_variable] = "/dev/null"
-            return saved_environ
+            return saved_environment_variables
 
-        def restore_environ(saved_environ: dict) -> None:
-            for saved_environ_key, saved_environ_value in saved_environ.items():
+        def restore_environ(saved_environment_variables: dict) -> None:
+            for saved_environ_key, saved_environ_value in saved_environment_variables.items():
                 if saved_environ_value is not None:
                     os.environ[saved_environ_key] = saved_environ_value
                 else:
@@ -111,7 +113,6 @@ class AwsContext:
         # Then our boto3 usage failed (here) with AWS credentials environment variables set.
         # Doing this just once (per AwsContext object creation) so as not to totally
         # undermine the (probably beneficial) caching that boto3 is trying to do.
-        #
         # Ref: https://stackoverflow.com/questions/36894947/boto3-uses-old-credentials
         # Ref: https://github.com/boto/boto3/issues/1574
         if self._reset_boto3_default_session:
@@ -119,6 +120,7 @@ class AwsContext:
             self._reset_boto3_default_session = False
         try:
             # Setup AWS environment variables for our specified credentials.
+            aws_credentials_dir = aws_credentials_dir_symlink_target = None
             if self._aws_access_key_id and self._aws_secret_access_key:
                 os.environ["AWS_ACCESS_KEY_ID"] = self._aws_access_key_id
                 os.environ["AWS_SECRET_ACCESS_KEY"] = self._aws_secret_access_key
@@ -132,14 +134,8 @@ class AwsContext:
                 if not os.path.isfile(aws_credentials_file):
                     raise Exception(f"AWS credentials file not found: {aws_credentials_file}")
                 os.environ["AWS_SHARED_CREDENTIALS_FILE"] = aws_credentials_file
-                if display:
-                    aws_credentials_dir_symlink_target = \
-                        os.readlink(aws_credentials_dir) if os.path.islink(aws_credentials_dir) else None
-                    if aws_credentials_dir_symlink_target:
-                        PRINT(f"Your AWS credentials directory (link): {aws_credentials_dir}@ ->")
-                        PRINT(f"Your AWS credentials directory (real): {aws_credentials_dir_symlink_target}")
-                    else:
-                        PRINT(f"Your AWS credentials directory: {aws_credentials_dir}")
+                aws_credentials_dir_symlink_target = (os.readlink(aws_credentials_dir)
+                                                      if os.path.islink(aws_credentials_dir) else None)
             else:
                 raise Exception(f"No AWS credentials specified.")
             if self._aws_region:
@@ -152,14 +148,30 @@ class AwsContext:
             # Setup AWS boto3 session/client to get basic AWS credentials info;
             session = boto3.session.Session()
             session_credentials = session.get_credentials()
+            if not session_credentials:
+                raise Exception("AWS session credentials cannot be determined.")
             caller_identity = boto3.client("sts").get_caller_identity()
+            if not session_credentials:
+                raise Exception("AWS caller identity cannot be determined.")
+            account_number = caller_identity["Account"]
+            user_arn = caller_identity["Arn"]
 
-            credentials = AwsContext.Credentials(session_credentials.access_key,
-                                                 session_credentials.secret_key,
-                                                 session.region_name,
-                                                 caller_identity["Account"],
-                                                 caller_identity["Arn"])
+            if not account_number:
+                raise Exception("AWS account number cannot be determined.")
+
+            credentials = AwsContext.Credentials(credentials_dir=aws_credentials_dir,
+                                                 credentials_dir_symlink_target=aws_credentials_dir_symlink_target,
+                                                 access_key_id=session_credentials.access_key,
+                                                 secret_access_key=session_credentials.secret_key,
+                                                 region=session.region_name,
+                                                 account_number=account_number,
+                                                 user_arn=user_arn)
             if display:
+                if aws_credentials_dir_symlink_target:
+                    PRINT(f"Your AWS credentials directory (link): {aws_credentials_dir}@ ->")
+                    PRINT(f"Your AWS credentials directory (real): {aws_credentials_dir_symlink_target}")
+                else:
+                    PRINT(f"Your AWS credentials directory: {aws_credentials_dir}")
                 PRINT(f"Your AWS access key: {credentials.access_key_id}")
                 PRINT(f"Your AWS access secret: {obfuscate(credentials.secret_access_key, show)}")
                 PRINT(f"Your AWS region: {credentials.region}")