+++ src/auto/utils/validate_utils.py
@@ -2,11 +2,12 @@ import os
 from typing import Optional
 from dcicutils.misc_utils import PRINT
 from .paths import (InfraDirectories, InfraFiles)
-from ..utils.misc_utils import (get_json_config_file_value,
-                                exit_with_no_action,
-                                print_exception)
+from ..utils.misc_utils import (
+    get_json_config_file_value,
+    exit_with_no_action,
+    print_exception,
+)
 from .aws import Aws
-from .aws_context import AwsContext
 
 
 def validate_and_get_custom_dir(custom_dir: str) -> (str, str):
@@ -16,12 +17,12 @@ def validate_and_get_custom_dir(custom_dir: str) -> (str, str):
     Exits with error if this value cannot be determined, or if the custom directory
     or JSON config file do not exist.
 
-    :param custom_dir: Explicitly specified path of the custom directory.
+    :param custom_dir: Explicitly specified path of the custom config directory.
     :return: Tuple with full paths of the custom directory and config file.
     """
     custom_dir = InfraDirectories.get_custom_dir(custom_dir)
     if not custom_dir:
-        exit_with_no_action("ERROR: Custom directory cannot be determined..")
+        exit_with_no_action("ERROR: Custom directory cannot be determined.")
     if not os.path.isdir(custom_dir):
         exit_with_no_action(f"ERROR: Custom directory does not exist: {custom_dir}")
     config_file = InfraFiles.get_config_file(custom_dir)
@@ -69,36 +70,70 @@ def validate_and_get_aws_credentials_dir(aws_credentials_dir: str, custom_dir: s
     return aws_credentials_dir
 
 
-def validate_and_get_aws_credentials(credentials_dir: str,
+def validate_and_get_aws_credentials(credentials_name: str,
+                                     credentials_dir: str,
+                                     custom_dir: str,
                                      access_key_id: str = None,
                                      secret_access_key: str = None,
                                      region: str = None,
                                      session_token: str = None,
-                                     show: bool = False) -> (Aws, AwsContext.Credentials):
+                                     show: bool = False) -> Aws:
     """
     Validates the given AWS credentials which can be either the path to the AWS credentials directory;
     or the AWS access key ID, secret access key, and region; or the AWS session token.
-    Prints the pertinent AWS credentials info, obfuscating senstive data unless show is True.
-    Returns a tuple with an Aws object and AwsContext.Credentials containing the credentials.
+    PRINTS the pertinent AWS credentials info, obfuscating senstive data unless show is True.
+    Returns an Aws object with these additional properites set:
+    - credentials: AwsContext.Credentials containing credentials which contains:
+      - credentials_dir: AWS credentials directory.
+      - credentials_dir_symlink_target: AWS credentials directory symlink target if credentials_dir is symlink.
+      - access_key_id: AWS Access key ID.
+      - secret_access_key: AWS Secret access key.
+      - region: AWS region name.
+      - account_number: AWS account number.
+      - user_arn: AWS account user ARN.
+    - credentials_name: AWS credentials name (e.g. cgap-supertest); this is set here not in AwsContext.
+    - custom_dir: Path of the custom config directory.
+    - custom_config_file: Path of the custom config file.
 
+    :param credentials_name: Explicitly specified AWS credentials name (e.g. cgap-supertest).
     :param credentials_dir: Explicitly specified path to AWS credentials directory.
+    :param custom_dir: Explicitly specified path of the custom config directory.
     :param access_key_id: Explicitly specified AWS access key ID.
     :param secret_access_key: Explicitly specified AWS secret access key.
     :param region: Explicitly specified AWS region.
     :param session_token: Explicitly specified AWS session token
     :param show: True to show any displayed sensitive values in plaintext.
-    :return: Tuple with an Aws object and AwsContext.Credentials containing the credentials.
+    :return: Aws object with credentials property set to AwsContext.Credentials containing credentials.
     """
+    custom_dir, custom_config_file = validate_and_get_custom_dir(custom_dir)
+    credentials_name = validate_and_get_aws_credentials_name(credentials_name, custom_config_file)
+    credentials_dir = validate_and_get_aws_credentials_dir(credentials_dir, custom_dir)
+
+    # Print header and basic info.
+    PRINT(f"Your custom directory: {custom_dir}")
+    PRINT(f"Your custom config file: {custom_config_file}")
+    PRINT(f"Your AWS credentials name: {credentials_name}")
+
     # Get AWS credentials context object.
     aws = Aws(credentials_dir, access_key_id, secret_access_key, region, session_token)
 
     # Verify the AWS credentials context and get the associated AWS credentials number.
     try:
         with aws.establish_credentials(display=True, show=show) as credentials:
-            return aws, credentials
+            if custom_config_file:
+                account_number_from_custom_config_file = get_json_config_file_value("account_number", custom_config_file)
+                if account_number_from_custom_config_file != credentials.account_number:
+                    exit_with_no_action(f"ERROR: AWS account number ({credentials.account_number}) does not match"
+                                        f" value ({account_number_from_custom_config_file})"
+                                        f" in config file: {custom_config_file}")
+            aws.credentials = credentials
+            aws.credentials_name = credentials_name
+            aws.custom_dir = custom_dir
+            aws.custom_config_file = custom_config_file
+            return aws
     except Exception as e:
-        print_exception(e)
         exit_with_no_action("ERROR: Cannot validate AWS credentials.")
+        print_exception(e)
 
 
 def validate_and_get_s3_encrypt_key_id(s3_encrypt_key_id: str, config_file: str, aws: Aws) -> Optional[str]: